# openapi/openapi.yml
openapi: 3.0.0
info:
  title: Horizon API
  version: 1.0.0
paths:
  /users/me:
    get:
      summary: Get current user profile
      tags:
        - Users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  responses:
    # Reusable responses to keep the paths clean
    Unauthorized:
      description: Authentication is required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: The requested resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

  schemas:
    BaseEntity:
      type: object
      required: [id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    User:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          properties:
            email: { type: string, format: email }

    # The RFC 9457 Compliant Schema
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          default: 'about:blank'
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code generated by the origin server.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
          description: Optional list of individual field errors (common for 400 errors).

    ValidationError:
      type: object
      properties:
        field:
          type: string
          description: The JSON path to the field that failed (e.g., "email").
        message:
          type: string
          description: A description of why the field failed validation.